pyproject.toml
[project]
name = "food-delivery-backend"
version = "0.1.0"
description = "Production-ready FastAPI backend for food delivery"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "aiosqlite>=0.20.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0"
]

[tool.ruff]
line-length = 100
target-version = "py313"
